requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.6
djangorestframework==3.15.1
sqlparse==0.5.0
tzdata==2024.1
django-allauth==0.62.1
requests==2.31.0
psycopg2==2.9.9
Pillow==10.2.0